==> 05-results.Rmd <==
# Results

```{r}
library(tidyverse)
library(patchwork)
library(readr)
library(ggplot2)
```

```{r}
job <- read_csv("finalJobs.csv")
```

## Most major job category

To help job seekers realize how many job postings are available for each job category and the competitiveness of various categories of jobs so that they can choose a more suitable job, We want to know the demand for different categories of jobs in New York city. From this job posting data variable "Job Category", for each job posting, our group extract the job category that can best represent the job, and store them and their counts in a new data frame "category".

```{r}
categorytrans <-job%>%select(`Job Category`, `Job ID`) %>% mutate(`Job Category` = as.character(`Job Category`),`Job Category` = gsub("([A-Za-z]+).*", "\\1", job$`Job Category`))

categorytrans$'Job Category'[which(categorytrans$`Job Category` == "Constituent")] <- "Constituent Services"

category <-
  as.data.frame(categorytrans["Job Category"]) %>%
  filter(!is.na(`Job Category`)) %>%
  group_by(`Job Category`) %>%
  summarise(count = n()) %>%
  arrange(desc(count))
  
ggplot(category, aes(x = fct_reorder(`Job Category`,count), y = count)) +
  geom_col(color = "white", fill = "dodgerblue") +
  ggtitle("Most Major Job Category") +
  labs(x = "Category", y = "Count") + 
  theme(plot.title = element_text(hjust = 0.5)) + 
  coord_flip()
```

From the graph above, we can find that the number of job posting for "Technology" is largest, which is more than 450, and the number of job posting for "Clerical" is smallest. For "Technology", "Constituent Services", and "Engineering", these three job categories have the top three number of job postings. If people want to choose jobs from these categories, they will have more choices than people who choose from "Social", "Information", and "Clerical".

```{r}
job$category = categorytrans$`Job Category`
job=job%>%filter(!is.na(category))
```

## Relationship of Salary & Category and Salary & Degree Requirement

For people who are finding jobs, the number of salary is always one of the most important information they are interested in. To make our graphs more useful and not to disappoint the job seekers, we choose the lower bound of salary range from each job posting whose salary frequency is annual as the value of salary. Then we create the box plots related to salary & job category and related to salary & degree requirement.

```{r}
Annualjob=job[job$`Salary Frequency`=="Annual",]
ggplot(Annualjob,aes(x = reorder(category,`Salary Range From`,FUN=median), y = `Salary Range From`/1000)) +
geom_boxplot( fill = "dodgerblue") +
ggtitle("Min Annual Salaries by Job Categories") + 
labs(x = "Job Category", y = "Min Annual Salary(unit: 1,000 usd)") + 
ylim(c(0,200))+
theme(plot.title = element_text(hjust = 0.5)) +
  coord_flip()
```

From the graph above, we can find except job categories "Clerical" and "Information", for each other job category, there are some outliers in its salary distribution. What's more, this graph shows that for those job postings whose job categories are "Technology" and "Information", the amount of salary people can get is generally higher than those job postings whose job categories are "Social" and "Clerical". In particular, for job category "Clerical", the amount of salary people can get is generally lower than any other categories of job. 

Combining with the first graph "Most Major Job Category" in section 5.1, we can notice that job category "Technology" not only has the largest variety of job postings, but also has better salary distribution than other job categories. For job category "Clerical", it has the smallest variety of job postings and the almost worst salary distribution. However, more variety of job postings does not always equal to better salary. For job category "Information", although it has the second lowest number of job postings, it ranks second in term of salary. In contrast, for job category "Constituent Services", although it ranks second in the number of job postings, it has the third lowest salary.

```{r}
Annualjob=Annualjob%>%filter(!is.na(degreereq))
ggplot(Annualjob,aes(x = reorder(degreereq,`Salary Range From`,FUN=median), y = `Salary Range From`/1000)) +
geom_boxplot( fill = "dodgerblue") +
ggtitle("Min Annual Salaries by degree") + 
labs(x = "Degree requirement", y = "Min Annual Salary(unit: 1,000 usd)") + 
ylim(c(0,200))+
theme(plot.title = element_text(hjust = 0.5)) +
  coord_flip()

```

From this graph, it shows a trend that except "Doctorate" (since the number of job postings whose degree requirement needs "Doctorate" is very small), the higher the degree requirement, the higher the salary. For job postings that need master's degree, the median of minimum annual salary is around 75000. But for those job postings that need associate degree, the median of minimum annual salary is just around 50000. Therefore, it is always beneficial for job seekers to improve their education level before looking for a job. 

```{r}
Annualjob$averageSalary=(Annualjob$`Salary Range To`+Annualjob$`Salary Range From`)/2
linerange<-Annualjob%>%select(`category`, averageSalary) %>%
  drop_na(averageSalary) %>%group_by(`category`) %>%summarise(meanvalue= mean(averageSalary)) 

ggplot() + 
  geom_linerange(data=Annualjob, aes(x=reorder(category,-averageSalary,FUN=mean), ymin=`Salary Range From`/1000, ymax=`Salary Range To`/1000), width=0.2, size=1, color="dodgerblue") +  
  geom_point(data=linerange, aes(x=category, y=linerange$meanvalue/1000), size=4, shape=21, fill="white")+
  theme(axis.text.x=element_text(angle=30))+
  ggtitle('Annual Salary range by Job category')+
  xlab('Job Category')+
  ylab('Salary Range(unit: 1,000 usd)')
```

Sometimes it is not enough for job seekers to just know the information of minimum salary. Here we plot this graph to show the mean salary, the maximum salary, and the minimum salary job seekers can get from each job category. Combining with the first graph "Min Annual Salaries by Job Categories" in section 5.2, for job categories "Engineering", "Administration", "Constituent Services", "Building", "Finance", and "Social", they rank higher based on mean salary than they rank based on median of minimum salary. However, for job category "Legal", although it ranks third in terms of median minimum salary, it has the second lowest number of mean salary, and the value of its mean salary is close to its median of minimum salary. This information shows job seekers who want to work in "Legal" job category can get a better minimum salary guarantee, but the salary they eventually get will not change much.

## Residency requirement and Full-Time / Part-Time job indicator

Although this dataset contains the job postings available in New York City, some job postings are not available for all job seekers in New York City since there exists some job postings that need New York city residency requirement. Furthermore, some job postings are not suitable for job seekers who just want to find a full-time job or just want to get a part-time job. In this section, we will analyze the residency requirement and the full-time / part-time indicator in these job postings. We group all job postings according to their job categories.

```{r}
cleveland2 <- job%>%select(`category`, `Job ID`,`residency`) %>%
  drop_na(`residency`) %>%group_by(`category`,`residency`) %>%
  summarise(count = n()) 

cleveland2<-cleveland2%>%pivot_wider(names_from = `residency`, values_from = count) %>%
  mutate(Total=`required`+`not required`) %>%
  ungroup() %>%
  pivot_longer(cols=!`category`, names_to="type", values_to="count") 
cleveland2[is.na(cleveland2)] <- 0
cleveland2$count[which(cleveland2$type == "Total"&cleveland2$category == 'Information')] <- 17
cleveland2$count[which(cleveland2$type == "Total"&cleveland2$category == 'Clerical')] <- 10
ggplot(cleveland2,aes(x=count,y=fct_reorder2(`category`,type== "Total",count,.desc=FALSE),color = type)) + 
  geom_point() +  
  ggtitle("Residency Requirement for Job Category") +
  ylab("") +
  theme(plot.title = element_text(hjust = 0.5))

```

In this graph, for each job category, it shows the amount of job postings that require the residency requirement and the amount that do not require, and the total amount. We can find for job category "Information", all of its job postings do not have residency requirement, and for job category "Clerical", all of its job postings need residency requirement. What's more, we notice that for most of the job categories, the amount of "required" is greater than the amount of "not required". Only in job categories "Technology", "Engineering", and "Information", the amount of job postings that do not require New York residency requirement is greater than the amount that require. This information shows now New York City provide a convenient employment environment for some international students whose major is about "Technology" or "Engineering". 

```{r}
cfjob = job %>% filter(!is.na(`Full-Time/Part-Time indicator`)) 

categ1 = cfjob %>% filter(`Full-Time/Part-Time indicator` %in% c("F", "P")) %>% filter(!is.na(category)) %>% count(category, `Full-Time/Part-Time indicator`)

total_categ1 = cfjob %>% filter(`Full-Time/Part-Time indicator` %in% c("F", "P")) %>% filter(!is.na(category)) %>% count(category)

most_categ1 = total_categ1[order(total_categ1$n, decreasing = TRUE),]

sort_categ1 = categ1 %>% filter(category %in% most_categ1$category)

ggplot() + geom_point(data = sort_categ1, aes(x=n, y=fct_reorder2(`category`,`Full-Time/Part-Time indicator`== "F",n,.desc=FALSE)
, color = `Full-Time/Part-Time indicator`)) + ggtitle("category and Full time / Part time job") + ylab("")
```

In this graph, we show the amount of full-time jobs and the amount of part-time jobs for each job category. It shows for job categories "Information" and "Clerical", they just have full-time jobs. For all job categories, the amount of full-time jobs are greater than the amount of part-time job. It means that full-time work seems to have gone mainstream in New York. What's more, we can find job category "Health" has the most number of part-time jobs. Thus, if you prefer to find a part-time in New York, "Health" category can provide more choices to you.

## Degree requirement and Career Level

In this section, we focus on analyzing the two most important requirements that job seekers will pay attention to when they look through these job postings: Degree requirement and Career Level. We want to know whether a higher career level can substitute for higher degree level?

```{r}
jobs=job%>% mutate(llevel = case_when(
    str_detect(`Career Level`, "Manager") ~ "Manager",
    str_detect(`Career Level`, "Entry-Level") ~ "Entry-Level",
    str_detect(`Career Level`, "Executive") ~ "Executive",
    str_detect(`Career Level`, "Experienced") ~ "Experienced",
    ))
jobs=jobs%>%filter(!is.na(llevel)) %>%filter(!is.na(degreereq))
```

```{r fig3, fig.height = 5, fig.width = 10,message=FALSE}
library(RColorBrewer)
library(ggmosaic)
jobs$degreereq = fct_relevel(jobs$degreereq, "high school diploma", "associate degree", "bachelor degree","master's degree","Doctorate")
jobs$llevel = fct_relevel(jobs$llevel, "Student", "Entry-Level", "Experienced","Manager","Executive")

ggplot(data = jobs) +
  geom_mosaic(aes(x = product(degreereq,llevel),fill=degreereq)) + 
  labs(title='')+ scale_fill_brewer(palette = "Blues")+xlab('Career Level')+ylab('Degree Requirement')
```

In this mosaic graph, it shows when career level changes from "entry-level" to "executive" (from low level to high level), the percent of high school diploma requirement become smaller and smaller, and the percent of master's degree requirement become larger and larger. The percent of Doctorate requirement is very low for all of these career levels. Also, the percents of bachelor degree requirement and associate degree requirement are similar in career levels "entry-level", "Experienced", and "Manager". But in career level "executive", the percents of both degree requirement are almost 0, almost all degree requirements are from master's degree and bachelor degree. Adding all these information together, we find higher career level cannot substitute for higher degree level, instead, when job postings' career level become higher, their degree requirements also become higher and do not need people whose degree is low.

```{r}

jobs=jobs%>% mutate(posttype = case_when(
    str_detect(`Posting Type`, "External") ~ "External",
    str_detect(`Posting Type`, "Internal") ~ "Internal",
    ))
jobs=jobs%>%filter(!is.na(llevel)) %>%filter(!is.na(degreereq))%>%filter(!is.na(residency))
ggplot(data = jobs) +
  geom_mosaic(aes(x = product(degreereq,posttype),fill=degreereq)) + 
  labs(title='')+ scale_fill_brewer(palette = "")+xlab('Posting Type')+ylab('Degree Requirement')+
  theme(axis.text.x=element_text(angle=30))+
  facet_grid(. ~residency)+
  ggtitle('Degree requirement ~ Posting type facet on Residency Requirement')
```

In this mosaic graph, we analyze the relationship between degree requirement and posting type & residency requirement.

1. For all job postings that do not require New York residency requirement: the degree requirement of external posting type is very similar to the degree requirement of internal posting type. The percent of bachelor degree requirement is more than 50% in both of these two posting type, which is the largest. The second largest percent of degree requirement is master's degree, which is around 33%. The percent of Doctorate requirement is smallest.

2. For all job postings that requires New York residency requirement: The percents of Doctorate and master's degree requirement are similar in both external and internal posting type. However, for job postings whose posting type is external, their percent of bachelor degree requirement is higher , and their percents of associate degree and high school diploma are smaller than those whose posting type is internal. 

3. For all external job postings: we can find the percents of master's degree and Doctorate degree requirement in "not required" are greater than those in "required". This shows the external job postings that do not need New York residency have higher requirements for high-end degrees than external job postings that need New York residency.

4. Similar for internal job postings, the internal job postings that do not need New York residency have higher requirements for high-end degrees than internal job postings that need New York residency.

## Title Classification
All NYC job positions have 4 job classes: competitive, non-competitive, exempt and labor.
Competitive means these positions require examination.
Exempt means competitive or non-competitive examinations or other qualification requirements are not practical.
Labor Positions are positions is mainly engaged in unskilled manual labor, like cleaner and driver.
Non-Competitive Positions are positions that are not in the exempt class or the labor class and examination has been determined to not be practicable. 
Is there any difference in salary between each class?


```{r}
D3jobs = job %>% mutate(minSalary = case_when(job$`Salary Frequency`=="Annual" & job$`Salary Range From` <= 25000 ~ "0-25k", job$`Salary Frequency`=="Annual" & job$`Salary Range From` <= 50000 & job$`Salary Range From` > 25000 ~ "25k-50k", job$`Salary Frequency`=="Annual" & job$`Salary Range From` <= 75000 & job$`Salary Range From` > 50000 ~ "50k-75k", job$`Salary Frequency`=="Annual" & job$`Salary Range From` <= 100000 & job$`Salary Range From` > 75000 ~ "75k-100k", job$`Salary Frequency`=="Annual" & job$`Salary Range From` > 100000 ~ ">100k"))

D3jobs = D3jobs %>% filter(!is.na(minSalary))
# write.csv(x = D3jobs, file = "D3jobs.csv", row.names=FALSE)
```

```{r}
library(ggridges)
job$`Title Classification` = fct_relevel(job$`Title Classification`, "Competitive-1", "Pending Classification-2","Labor-3","Exempt-4","Non-competitive-5")
ggplot(job, aes(x=`Salary Range From` / 1000, y=`Title Classification`)) + geom_density_ridges(alpha=0.6) + xlab("salary range from (unit: 1000 dollar)")
```

In this ridgeline plots, one for each kind of "Title Classification", it shows us how the number of job postings of each title classification changes with the lower bound of salary range. We can find that for job postings whose title classifications are "Competitive-1", "Pending Classification-2", and "Non-competitive-5", their low bound of salary range are concentrated around 50,000 dollars. "Exempt-4" is more likely to get a higher salary, with one peak near 100,000 dollars and one peak near 150,000 dollars. However, for job postings whose title classification is "Labor-3", most of it low bound salary range is less than 50,000 dollars, which may because it is unskilled job.

## External and Internal

Except the New York residency requirement, these job postings have another restriction variable "Posting Type" whose meaning is internal or external job posting. And the dataset claims that internal postings available to city employees and external postings available to the general public are included. To find more information about internal and external, we create a scatter plot to show the relationship between job category and this posting type.

```{r}
scatter=job%>%select(`category`, `Job ID`,`Posting Type`) %>%
  drop_na(`Posting Type`) %>%group_by(`category`,`Posting Type`) %>%
  summarise(count = n()) %>%pivot_wider(names_from = `Posting Type`, values_from = count) 

scatter<- scatter %>% 
  mutate(type = case_when(External>Internal*0.8~ "External ",
                          External<Internal*0.6 ~ "Internal",
                          External>Internal*0.6&External<Internal*0.8 ~ "Neutral"))
library(ggrepel)

scatter[is.na(scatter)] <- 0
ggplot(scatter,aes(External, Internal,color=type)) + geom_point()+geom_label_repel(aes(label = scatter$category), size = 3) +
  xlab("External count") +
  ylab("Internal count")
```

In this graph, for each job category, we set some categories are more "external" if its number of external job postings is greater than the number of internal job postings times 0.8, some categories are more "internal" if its number of external job postings is less than the number of internal job postings times 0.6, and else categories are "neutral".
From the graph, we can find for job categories "Clerical", "Social", and "Administration", these categories are more "internal". For job categories "Information", "Policy", "Public", "Legal", and "Constituent Services", they are more "external". Other categories are "neutral".  Thus, for job seekers who are not city employees now, it is difficult to work in Administration. 

## Word Cloud for Preferred Skills

We noticed that each job has a specific long paragraph to describe the Preferred Skills,  so we will focus on the keywords that appear in  Preferred Skills from the whole job posting dataset.  By split all words and drop all meaningfuless word like conjunctions, numbers and symbols, we generate a worldcloud with 
100 most frequent words. 

```{r}


library(dplyr); 
library(tidytext); 
wordJobs<-job

wordc<-wordJobs%>%
  unnest_tokens(input = `Preferred Skills`, output = word)%>%
  select(word)%>%
  anti_join(stop_words)%>%
  group_by(word)%>%
  summarize(freq = n())%>%
  ungroup()%>%
  arrange(desc(freq))%>%
  top_n(100)

wordc<-wordc%>%
  filter(word!='e9')
wordc<-wordc%>%
  filter(word!='1', word!='0',word!='2'
         ,word!='3',word!='4',word!='5'
         ,word!='6',word!='7',word!='8'
         ,word!='9',word!='10',word!='0080',word!='00a1',word!='0099')



library(wordcloud2)
#wordc include word and word frequency
wordcloud2(wordc,size = 2)
 
```
There are many information we can extract from above wordcloud. For example, we can see that'Data','Microsoft','excel' appear around 1000 times, and 'team','project', 'manage' appear more than 500 times. This may indicate most employer expect the job seeker are familar with microsoft office and have basic data analysis ability. What's more, experience with team project may also be a bonus for job seeker.
