==> 06-interactive.Rmd <==
# Interactive component



## D3 interactive Annual Salary Distribution

For job seekers who want to work in NYC but do not know about NYC salary level, this interactive part can help them to get a salary expectation.
For job seekers who have a salary expectation but not sure apply for which category, this interactive part can help them to compare each category.


Click the button at the bottom to select the job category or view all you are interested in.
If you want to know the specific quantity, you can hover the mouse on each bar.




<script src="https://d3js.org/d3.v6.min.js"></script>
<iframe src="https://vizhub.com/wyx34/1f5e5085ddfd4881a6d6e271cca0afca?mode=embed" title="Annual Salary Visualization" height="650" width="800"></iframe>

To view standalone version, Please click : [Interactive-annual-salary.html](Interactive-annual-salary.html)

## Prefered skills of Engineer 
Use wordcloud2 package to build this wordcloud.
If you want to know the specific quantity, you can hover the mouse on each word.
```{r}
library(tidyverse)
library(patchwork)
library(readr)
library(ggplot2)
```


```{r}
library(dplyr); 
library(tidytext);
library(wordcloud2)

D3jobs<-read_csv("D3jobs.csv")

wordjob2<-D3jobs%>%
  filter(category=='Engineering')


word1<-wordjob2%>%
  unnest_tokens(input = `Preferred Skills`, output = word)%>%
  select(word)%>%
  anti_join(stop_words)%>%
  group_by(word)%>%
  summarize(freq = n())%>%
  ungroup()%>%
  arrange(desc(freq))%>%
  top_n(100)


word2<-word1%>%
  filter(word!='e9', word!='0080',word!='00a1',word!='0099',word!='skills')
word3<-word2%>%
  filter(word!='1', word!='0',word!='2'
         ,word!='3',word!='4',word!='5'
         ,word!='6',word!='7',word!='8'
         ,word!='9',word!='10')
word3

wordcloud2(word3,size = 2)
```
