==> 07-conclusion.Rmd <==
# Conclusion


From our project, we can find the jobs in technology are needed most and jobs in Clerical are needed least in New York City. If job seekers want to get a higher salary, it will be good choice for them to get a master's degree and then work in the technology or information field. What's more, higher career level cannot substitute for higher degree level, it is always beneficial for job seekers to improve their education level before looking for a job. From the Word Cloud, we can notice that if job seekers want to improve their competitiveness, it would be better for them to improve their skills about Excal, Microsoft, and project experience. For people who have no direction but have specific requirements for the future salary range, for example, they want to achieve an annual salary between 75000 and 100000 dollar, it would be better for them to find jobs whose title Classification is "Exempt-4" or "Non-competitive-5", and technology kind of job is the most of recommended job for them, followed by engineering kind of job. Most of job postings for these two kinds of job are neutral type, full time job and do not need New York City residency. After getting these information, these job seekers can better plan for their future. 



From this project we learned that data processing skills and extract information ability are the foundation of data visualization. For each visualization, we spent more time on chose the right type and deal with the input data than draw the plot.
The limitations in our project are that the objects we analyze and conclude are just a series of job postings that have some same information. We did not analyze some specific jobs or return some specific Job IDs to help job seekers find their job. The future direction for our project is to figure out these limitations and try to help job seekers find the most suitable job with its specific job ID, not just a group of jobs.
